adc works. both pump and fan sense channels. Note: 3v3 is max adc reading

+++ common/src/packet.rs
@@ -33,10 +33,10 @@ pub struct AcceptConnectionPacket {
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
 pub struct ReportSensorsPacket {
     /// Normalized representation of the fan's rpm.
-    pub fan_speed_norm: u8,
+    pub fan_speed_norm: u16,
 
     /// Normalized representation of the pump's rpm.
-    pub pump_speed_norm: u8,
+    pub pump_speed_norm: u16,
 
     /// Valve State
     pub valve_state: ValveState,
@@ -77,11 +77,11 @@ impl Into<bool> for ValveState {
 pub struct ReportControlTargetsPacket {
     /// The voltage value which the embedded hardware should immediately output
     /// for the fan.
-    pub fan_control_voltage: u8,
+    pub fan_control_voltage: u32,
 
     /// The voltage value which the embedded hardware should immediately output
     /// for the pump.
-    pub pump_control_voltage: u8,
+    pub pump_control_voltage: u32,
 
     /// The valve is either instructed to begin opening or closing.
     /// Sending the state which the valve is in results in nothing happening.
